Sixth commit

+++ src/main/java/pl/blazejherzog/mywallet/transactions/TransactionRepository.java
@@ -2,5 +2,5 @@ package pl.blazejherzog.mywallet.transactions;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 }
